+++ ProjectileDrop.py
@@ -17,61 +17,7 @@ Assumptions:
 """
 
 import math
-
-"""
-The point class specifies a point in three dimensions
-    @param xComp The X Component running laterally to the ground
-    @param xComp The Y Component perpendicular to the ground. The ground is presumed to be 0
-    @param xComp The Z Component longitudinal to the ground
-
-"""
-class point:
-    def __init__(self, xComp: float, yComp: float, zComp: float):
-        self.x = xComp
-        self.y = yComp
-        self.z = zComp
-    def __str__(self):
-        return ("X: "  + str(self.x) + " Y: " + str(self.y) + " Z: " + str(self.z))
-
-    def getX(self):
-        return self.x
-
-    def getY(self):
-        return self.y
-
-    def getZ(self):
-        return self.z
-
-    def setX(self, newX: float):
-        self.x = newX
-
-    def setY(self, newY: float):
-        self.y = newY
-
-    def setZ(self, newZ: float):
-        self.z = newZ
-
-"""
-The vector class specifies a vector in three dimensions
-    @param xComp The X Component running laterally to the ground
-    @param xComp The Y Component perpendicular to the ground. The ground is presumed to be 0
-    @param xComp The Z Component longitudinal to the ground
-
-"""
-class vector:
-    def __init__(self, xComp: float, yComp: float, zComp: float):
-        self.x = xComp
-        self.y = yComp
-        self.z = zComp
-
-    def getX(self):
-        return self.x
-
-    def getY(self):
-        return self.y
-
-    def getZ(self):
-        return self.z
+from ProjectileLocation import *
 
 """
 The dropCalclations class specifices the location and time? when a projectile should dropped to land on a target
